requirements.txt
python == 3.8.13
pytorch == 1.12.1
torchvision == 0.13.1
transformers == 4.30.2
pandas == 1.4.3